admin_all_routes_working

+++ src/Components/Pages/Auth/RegisterFrom.jsx
@@ -7,6 +7,7 @@ import logoWhite from "../../../assets/images/logo/logo.png";
 import logoDark from "../../../assets/images/logo/logo_dark.png";
 import { useNavigate } from "react-router-dom";
 import { createAccount } from "../../../api/integrateConfig";
+import Swal from 'sweetalert2';
 
 
 const RegisterFrom = ({ logoClassMain }) => {
@@ -22,8 +23,8 @@ const RegisterFrom = ({ logoClassMain }) => {
     email: '',
     password: '',
     confirmPassword: '',
-    referralID: '',
-    referralName: '',
+    referById: '',
+    referByName: '',
     address: '',
     referBy: '',
   });
@@ -56,17 +57,29 @@ const handleSubmit = async (e)=>{
   e.preventDefault();
   console.log(`formdata is :`)
   console.log(formData)
+  if(formData.password !== formData.confirmPassword){
+    alert("passwords do not match")
+    return ;
+  }
   // const response = await createAccounts
   try{
   const response = await createAccount(formData);
-  console.log(`the response recieved from response is : ${response.message}`);
+  // console.log(`the response recieved from response is : ${response.message}`);
   console.log(response)
-  localStorage.setItem("address" , formData.address)
-  localStorage.setItem("userID" , response.userId)
+  localStorage.setItem("login", JSON.stringify(true));
+  localStorage.setItem("authToken" , response.token);
+  localStorage.setItem("address" , response.address);
+  localStorage.setItem("userID" , response.userId);
+  Swal.fire({
+    icon:"success",
+    title:"SUCCESSFULL",
+    text:"Successfully created profile! ",
+  })
   }catch(error){
     console.log(error)
+    alert("Please login!")
   }
-  // navigate(`${process.env.PUBLIC_URL}/dashboard/default/`);
+  // navigate(`/dashboard/default/`);
 
 
 }
@@ -201,7 +214,7 @@ const handleSubmit = async (e)=>{
                   <Input
                     className="form-control"
                     type="name"
-                    name="referralID"
+                    name="referById"
                     required
                     defaultValue={ref}
                     onChange={handleChange}
@@ -217,7 +230,7 @@ const handleSubmit = async (e)=>{
                   <Input
                     className="form-control"
                     type="text"
-                    name="referralName"
+                    name="referByName"
                     onChange={handleChange}
 
                   />
@@ -277,7 +290,7 @@ const handleSubmit = async (e)=>{
                 Already have an account?
                 <Link
                   className="ms-2"
-                  to={`${process.env.PUBLIC_URL}/login`}
+                  to={`/login`}
                 >
                   Sign in
                 </Link>
